+++ changelog.md
@@ -10,6 +10,7 @@ Saves from 2.5 are not compatible with 3.0.
 * **[Campaign AI]** Every 30 minutes the AI will plan a CAP, so players can customize their mission better.
 * **[Campaign AI]** AI now considers Ju-88s for CAS, strike, and DEAD missions.
 * **[Campaign AI]** Fix purchase of aircraft by priority (the faction's list was being used as the priority list rather than the game's).
+* **[Flight Planner]** AI strike flight plans now include the correct target actions for building groups.
 * **[UI]** Added new web based map UI. This is mostly functional but many of the old display options are a WIP. Revert to the old map with --old-map.
 * **[UI]** Campaigns generated for an older or newer version of the game will now be marked as incompatible. They can still be played, but bugs may be present.
 * **[UI]** DCS loadouts are now selectable in the loadout setup menu.